pyproject.toml
[project]
name = "roadauth-pro"
version = "0.1.0"
description = "BlackRoad Enterprise Authentication - OAuth2, JWT, MFA, SSO"
authors = [{name = "BlackRoad OS", email = "blackroad.systems@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
keywords = ["auth", "oauth2", "jwt", "mfa", "sso", "enterprise", "blackroad"]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    "httpx>=0.26.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0", "black>=24.0.0", "ruff>=0.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roadauth"]

[project.scripts]
roadauth = "roadauth.main:cli"
